usePlayer: extract sleep-timer state machine to useSleepTimer hook
Sleep-timer logic was three refs + a state + two callbacks living
inside `usePlayer` for no good reason — none of it touches anything
else in the player except `endOfTrackPendingRef` (the flag
`advanceRef` reads to suppress the next natural advance). Moved
into its own hook in `useSleepTimer.ts`; the player passes the
ref down so the "stop at end of track" mode can still flip it.

Now self-documented (the modes — minutes / end-of-track / null —
each have their own paragraph in the docstring) and could be
unit-tested in isolation if we ever want to.

usePlayer.ts: 1257 → 1217 LOC. tsc clean. vitest 36 passed.
